Update tests and resolve test failures
